fix(stepfunctions-tasks): checking for task token in EcsRunTask containerOverrides causes memory explosion (#15187)
EcsRunTask containerOverrides had references to the entire construct tree causing a large amount of recursion
when trying to validate that the task token was provided in the waitForTaskToken integration pattern. The only place for the task token to appear is in the environment of a containerOverride, so we should only check for it there.

I also updated the error to mention JsonPath instead of Context since Context is deprecated.

In addition to the unit tests, I packaged the library to dotnet and succesfully synthed the minimal reproduction.

thanks to @BenChaimberg for providing the suggested fix

fixes #15124

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*