Adding token check to input validation
We currently do not check if inputs are token when validating them which might throw errors
if it is given a `ref`, adding them in here and corresponding unit tests that do not throw validation errors
when given tokens with invalid inputs.

Also updated one test which had mismatched description.

Testing done
------------------
* `yarn build && yarn lint & yarn test`