fix(servicecatalog): Portfolio fails validation when passed Tokens as its properties (#15208)
We currently do not check if inputs are token when validating them which might throw errors
if it is given a `ref`, adding them in here and corresponding unit tests that do not throw validation errors
when given tokens with invalid inputs.

Also updated one test which had mismatched description.

Also moved core to qualified import for better consistency. 

Testing done
------------------
* `yarn build && yarn lint & yarn test`


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*