fix(elasticsearch): Domain.fromDomainAttributes gives "Invalid URL" when endpoint is a token (#15219)
`Domain.fromDomainAttributes` throws the error "Invalid URL" when a token endpoint is provided. In this PR, the domain name is retrieved from the scope using the arn present in `DomainAttributes`, and if no name is found, we try to get it from the endpoint (as it has been done so far). Also added a test to verify that the error disappears with this change, even when it is given token attributes.

closes #15188

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*