chore: add `App.addCustomSynthesis`
In "legacy" constructs, constructs had the ability to have themselves
activated during synthesis, so they could add additional output
artifacts to the Cloud Assembly.

In #14974, we made good use of that to have the `PublishAssetAction`
have the ability to emit the buildspec it generated to a file
instead of to the CloudFormation template, improving pipeline
performance.

In constructs v10, and hence CDKv2, that ability got removed since the
Cloud Assembly and "synthesis" are domain-specific concepts that
aren't part of the constructs library.

This now breaks the merge to v2 in #15132, since `PublishAssetAction`
uses a concept that cannot be translated.

------------

Add an AWS CDK-specific mechanism back to have constructs inject
themselves into the synthesis lifecycle, and make `PublishAssetAction`
use that new mechanism.