docs(s3-deployment): note that asset inputs must come from a trusted source (#15276)
In #15244, it was brought to our attention that `zip.extractall()` in
Python is not safe against out-of-tree file writing attacks.

There is probably a way in which a clever attacker might use that to
execute arbitrary code in the Lambda Function, and read the bucket or
write file to it in unexpected places.

Add a note to the README indicating that this module should only
be used with Zips from trusted sources (like asset bundles built by
the CLI).

Not a full fix but a mitigation for #15244.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*