fix(autoscaling-common): create all scaling intervals correctly when minimum lower bound is not zero
Before this fix, the first scaling step would be ignored whenever its lower bound was greater than zero (see linked issue).

Now, instead of *replacing* the first interval with a "dummy" one, we *prepend* the list of intervals. This behavior complements what was already implemented in case the last interval's upper bound was not Infinity: a dummy one was appended.

This fixes #10141