fix(pipelines): artifact bucket permissions missing for in-account deployments
A recent change to the default `deploy-role` policies required that the
required permissions statements are added to the artifacts bucket
instead.

This happened to be working for stacks which had NO `account` env property
set, but not for stacks that DID have an `account` property set which
was the same as the pipeline's account property.

Fixes #15307.