feat(awsservicecatalogappregistry): Add association methods
This PR adds association methods for AppRegistry application construct.
Users can associate attribute groups and stacks to their application. A current limtitation
is that a stack can only be associated with one application, we would like to have ability to
fail faster (during `cdk synth`) if this logic is violated in an app instead of during `cdk deploy`
which it now currently does.

Testing done
------------------
* `yarn build && yarn test && yarn integ`
* `yarn rosetta:extract`

Co-authored-by: Dillon Ponzo <dponzo18@gmail.com>