feat(servicecatalogappregistry): allow Applications to associate with AttributeGroups and Stacks (#15371)
This PR adds association methods for AppRegistry application construct.
Users can associate attribute groups and stacks to their application. A current limtitation
is that a stack can only be associated with one application, we would like to have ability to
fail faster (during `cdk synth`) if this logic is violated in an app instead of during `cdk deploy`
which it now currently does.  That will be addressed in an upcoming PR if still relevant. 

Testing done
------------------
* `yarn build && yarn test && yarn integ`
* `yarn rosetta:extract`



----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*


Co-authored-by: Dillon Ponzo <dponzo18@gmail.com>