fix(pipelines): artifact bucket permissions missing for in-account deployments (#15348)
A recent change to the default `deploy-role` policies required that the
required permissions statements are added to the artifacts bucket
instead.

This happened to be working for stacks which had NO `account` env property
set, but not for stacks that DID have an `account` property set which
was the same as the pipeline's account property.

Ultimately this was caused by an account comparison deep in the bowels of the
Grants system, which happened to work out because in cases of mismatched
accounts it would think there was a cross-account access and it was trying to
add permissions to both principal and resource. But in this case, it's much more
accurate to say that we can NEVER add permissions to the `deploy-role`, and
instead always want the permissions added to the bucket (rather than dropping
them).

Fixes #15307.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*