fix(core): unresolved tokens in generated nested stack outputs (#15380)
When the CDK generates a CloudFormation output to communicate information from the nested stack, it uses the `reference.displayName` property to increase uniqueness. If the display name includes a token, the id of the generated output includes a token and that’s not allowed.

The fix is to do the same thing we do for parameter IDs: `resolve()` the generated output ID before using it.

Fixes #15155


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*