feat(pipelines): Docker registry credentials (#15364)
Introduce a new set of properties to the pipeline constructs to enable users to
specify Docker registries -- and associated credentials for each -- to be used
during the pipeline build/synth, self-mutate, and asset publishing stages.

These APIs enable the user to specify a Docker registry (e.g., DockerHub, ECR)
and either secrets or use role credentials to authenticate to each registry, as
well as specify which step(s) of the pipeline need these credentials.

fixes #10999
fixes #11774

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*