fix(pipelines): unable to add assets stage to existing VPC pipeline
The VPC Policy for the asset publishing roles is created as a separate policy,
which is guaranteed to be created after the role, but not necessarily before
it's needed. CodeBuild -- at CloudFormation deploy time -- actually verifies the
roles have appropriate permissions, leading to a potential race condition and
failure if VPC permissions are not present when the CodeBuild jobs are created.

Add an explicit dependency on the separate policy to guarantee creation order
during deployment.

fixes #14343