doc(pipeline): document work-arounds for when bootstrap stack has been recreated
If bootstrap stack roles have been deleted and recreated, it results in cryptic
errors in the pipeline (AccessDeniedException). An elegant and robust fix to
this is involved and not likely to be implemented in the near-term.

In the meantime, what we can do is document ways to work around this issue, and
heavily advise people not to get themselves into this scenario in the first
place. This README update intends to accomplish the former, if not the latter.

Tested and verified the proposed "automated" solution (re-creating bootstrap
stacks).

related #11934