feat(bootstrap): widen lookup role permissions for future extension
The lookup role permissions are currently scoped to support exactly
the context lookups supported by CDK right now.

This means that if we add more lookups in the future, we will have to
do a complicated dance that involves updating the bootstrap stack and
version checking.

Instead, increase the scope of the lookup role to be able to read
everything in the target account except for secrets (anything that's
encrypted by KMS cannot be read).

People that currently have built their own lookup extensions to CDK can
use this role to implement them.