feat(servicecatalog) Add portfolio-product association, TagOptions, and TagUpdate constraint
Adding ability to associate a product with a portfolio.
Adding TagOptions to a portfolio.
Adding initial framework for constraints, starting here with TagUpdate Constraints.  Since there
is considerable amount of code in the `association-manager` it felt more manageable to split these PRs
so that adding the other constraints only requires adding the APIs and then direct construction methods.

Testing done
------------------
* `yarn build && yarn test`
* `yarn integ`
* `yarn rosetta extract:strict`

Co-authored-by: Dillon Ponzo <dponzo18@gmail.com>:wq