feat(aws-elasticloadbalancingv2): Allow listing added listeners on application load balancers (#15259)
This PR adds the ability to get access to the list of Listeners that have been added to an Application Load Balancer.  This is necessary because when an L3 pattern creates a listener on a load balancer the caller may need to get access to the listener to customize it.

I couldn't figure out a great way to make this work for Looked Up or Imported Load Balancers so I decided to throw an error.  If there is a better way, let me know and I can change the PR. 

closes #11841
----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*