fix(autoscaling): scaling intervals are incorrect if the bottom one does not start at 0  (#15345)
Before this fix, the first scaling step would be ignored whenever its lower bound was greater than zero (see linked issue).

Now, instead of *replacing* the first interval with a "dummy" one, we *prepend* the list of intervals. This behavior complements what was already implemented in case the last interval's upper bound was not Infinity: a dummy one was appended.

This fixes #10141


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*