chore(cdk-release): do not get commits if Changelog generation is skipped (#15467)
Our build in the Pipeline bump the version to a prerelease one before running the integ tests.
We recently switched to using the `cdk-release` tool,
and it always fetches the Git history when doing the bump.
But that's a problem, as our Pipeline build does not run in a Git repository,
and querying the Git history there crashes.

Make it so that querying the Git history is skipped when Changelog generation is skipped
(as that's when those commits are actually used).

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*