fix(aws-ecs): token is added to Options instead of SecretOptions in SplunkLogDriver (#15408)
----
This PR closes [#7264](https://github.com/aws/aws-cdk/issues/7264).

The `token` field of the Splunk log driver populates the `Options` property of the Log Configuration which leads to the secret being resolved to its value on deploying, and then the token is viewable in plain text in the console and may be stored in plain text elsewhere. Thus, we are deprecating the `token` field of the Splunk log driver and are introducing a new `secretToken` field. `secretToken` can be used to provide the Splunk token as a Secrets Manager Secret or a Systems Manager Parameter and will be populated in the `SecretOptions` property of the Log Configuration. 

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*