feat(pipelines): revised version of the API (#12326)
Add a new, modernized API to the `pipelines` library.

Advantages of the new API are:

- Removes the need to interact with the underlying AWS CodePipeline library for `Artifacts` and `Sources`
- A streamlined API for sources (more sensible defaults allowing you to specify less)
- `Synth` classes hide less from you, allowing you more control and remove the need to decide whether or not to "eject" from the convenience classes of the original API
- Supports parallel deployments (speeding up large pipelines)
- Supports stages of >25 stacks
- Supports multiple sources powering the build
- Gives more control over the CodeBuild projects that get generated

In addition, by clearly separating out generic parts of the library from CodePipeline/CodeBuild-specific parts, allows easier development of construct libraries that target alternative deployment systems while reusing large parts of the logic of this library.

This does not remove or deprecate the old API, though starting today its use is discouraged in favor of the new API, which will see more development in the future.

Closes #10872.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*