fix(cfnspec): make EndpointConfiguration of AWS::Serverless::Api a union type (#15526)
A recent update to the SAM spec (https://github.com/aws/aws-cdk/pull/15311)
changed the EndpointConfiguration property of AWS::Serverless::Api to have a complex type.
However, that is a breaking change compared to the previous, string, type.
I consulted with the SAM team, and it turns out the property accepts both a string and the complex type.
Given that, patch our SAM spec to make EndpointConfiguration a union type.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*