chore(cli): tests that apply only to v1 are failing when building for v2 (#15565)
In v1 we are allowing selection by the ID of the stack, with a warning that it will be stop working in v2 and asking the user to use the same names shown in the `cdk ls` result. Since the tests are written with the ID as input, they work when building for v1, but fail when building for v2. This change makes it work in both cases.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*