chore(kms): cross-stack usage detection depends on NPM tree (#15580)
KMS keys try to be smart about not generating impossible dependencies
between multiple stacks, which CodePipeline takes advantage of for
its support stacks.

However, because the logic that tests for this case has an `instanceof
Construct` in its code path, if there are ever multiple copies of the
`constructs` library in the NPM tree the test will fail, and the
resulting error will be very confusing.

This situation can arise when people flip back and forth between
CDK v1 and v2, because `package-lock.json` will contain half-baked
dependency trees; people will be looking at their code but the issue
will be in invisible state.

Be more liberal in detecting that a construct is, in fact, a construct
to get around this.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*