feat(cli): add ability to specify an external id for the deploy-role (#15604)
added the ability to specify an external id for the deploy-role as part of the synthesizer config.
this is similar to the existing functionality that allows specifying the external id for the image
and file publishing roles

In order to take advantage of this functionality you would need to customize the bootstrap template. To test this feature I customized the DeploymentActionRole in the bootstrap template to have the `AssumeRolePolicyDocument` as:

```yaml
DeploymentActionRole:
    Type: AWS::IAM::Role
    Properties:
      AssumeRolePolicyDocument:
        Statement:
          - Action: sts:AssumeRole
            Condition:
              StringEquals:
                sts:ExternalId: "my-external-id"
            Effect: Allow
            Principal:
              AWS:
                - Fn::Join:
                    - ""
                    - - "arn:"
                      - Ref: AWS::Partition
                      - :iam::1111111111111:root
        Version: "2012-10-17"
      ...
      RoleName:
        Fn::Sub: cdk-${Qualifier}-deploy-role-${AWS::AccountId}-${AWS::Region}
```

I then created a test stack that specified that external id:

```ts
new CdkExternalIdTestStack(app, 'CdkExternalIdTestStack', {
  env: {
    account: '111111111111',
    region: 'us-east-2',
  },
  synthesizer: new cdk.DefaultStackSynthesizer({
    deployRoleExternalId: 'my-external-id',
  }),
});
```
----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*