chore: V2 version of the `@aws-cdk/assertions` library
The `@aws-cdk/assertions` library was previously excluded from the V2 build of the experimental modules,
because the `vendor.sh` script it used for vendoring in a few modules
(`cfnspec`, `cloudformation-diff`, and `assert-internal`)
used relative paths inside the repository.
Change the script to use relative paths instead,
and also adjust the module copying logic to account for bundled dependencies,
which `@aws-cdk/assertion` uses.