introduce pragma:context to cdk-integ
Since context passed via command line overrides the App context, setting newStyleSynthesis in the pipeline integration tests will not work since in the v2 branch we set newStyleSynth to “false” explicitly in cdk-integ.

Introduce a new cdk-integ pragma that can be used to override this context and use it for pipelines integration tests.