chore: some cli tests still rely on legacy behavior and fail in v2
A similar case to #15565 in which CLI tests assume stacks can be selected by ID and then fail when they land in v2-main. 

To detect this upstream (on `master`), I added an environment variable that explicitly disables the legacy behavior in cx-api and use it in the CLI tests to ensure that tests are executed without legacy behavior enabled.