chore: some cli tests still rely on legacy behavior and fail in v2 (#15633)
A similar case to #15565 in which CLI tests assume stacks can be selected by ID and then fail when they land in v2-main. 

To detect this upstream (on `master`), I added an environment variable that explicitly disables the legacy behavior in cx-api and use it in the CLI tests to ensure that tests are executed without legacy behavior enabled.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*