feat(assertions): retrieve matching resources from the template
Provide an API `getResources()` that retrieves the matching resources
from the template given its type and optional predicate.

For complex assertions that cannot be modeled using the primitives
provided by this module, this API allows an 'escape hatch' so that
assertions can be written directly into the test case.

This is being used in the `aws-cloudwatch` module, specifically to
assert widgets in a CloudWatch Dashboard that are modeled as serialized
JSON within a property in the resource.