chore: fix versions of inner dependencies in @aws-cdk-lib (#15648)
The build of the V2 experimental modules hardcoded the versions of unstable-to-unstable dependencies to `0.0.0`,
which does not work in our pipeline, which first runs the `align-versions.sh` script to give all modules a temporary version.
Fix it by using the version the experimental package itself is (as that's the version assigned by `align-versions.sh`).

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*