fix(events): cross-account event targets that have a Role are broken
When we started supporting cross-region event targets, we needed
to add a role to the event-bus target. At that point, we also
opted to fall back to the role that the event target requested
for itself.

However, that was wrong: the role used in that place is *only*
for passing events between event buses, and *never* for triggering
the actual target.

Solution: don't fall back, only use a special role for event passing.
In fact, don't even do it if the target isn't in a different region,
because apparently it's not necessary for cross-account event
passing at all.

Refactor the `events` code a little to make clear what is happening
and why, because it was string to get messy.

Fixes #15639.