chore: introduce pragma:context to cdk-integ (#15779)
Since context passed via command line overrides the App context, setting newStyleSynthesis in the pipeline integration tests will not work since in the v2 branch we set newStyleSynth to “false” explicitly in cdk-integ.

Introduce a new cdk-integ pragma that can be used to override this context and use it for pipelines integration tests.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*