chore(assertions): rename TemplateAssertions to Template
The class `TemplateAssertions` is already part of the
`@aws-cdk/assertions` module. The word 'assertions' is now redundant.

Further, this class now has the method `findResources()` which is not an
assertion anymore.
All assertion methods are named in a way that makes it clear that it is
an assertion, viz., `hasResource()`, `resourceCountIs()`,
`templateMatches()`.