chore(assertions): rename TemplateAssertions to Template (#15823)
The class `TemplateAssertions` is already part of the
`@aws-cdk/assertions` module. The word 'assertions' is redundant.

Further, this class now has the method `findResources()` which is not an
assertion.
All assertion methods are named in a way that makes it clear that it is
an assertion, viz., `hasResource()`, `resourceCountIs()`,
`templateMatches()`.

BREAKING CHANGE: `TemplateAssertions` is now renamed to
`Template`.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*