chore(ubergen): strip alpha from experimental module names
As part of #15722, the jsii targets for experimental packages on the v2 branch
were changed to include an 'alpha' identifier so the package and module names
would more obviously reflect the stability of the packages.

However, since this change was made directly to the package.json files, this
means that all usages of these targets across the source were impacted,
including usages via ubergen. The impact of this is that the L1s published to
aws-cdk-lib were incorrectly published under this 'alpha' module instead of
their normal location. L1s should still go under the standard location, as they
are stable; only the experimental L2s should be impacted by the alpha
designator.

fixes #15845