chore: rewrite jsii targets for experimental packages during transformation
As part of #15722, the decision was made to change the package.json files for
experimental packages to add 'alpha' identifiers. This ends up causing issues
for our ubergen-ed packages (i.e., aws-cdk-lib), which then try to include L1s
for these packages under the 'alpha' namespaces. Instead, switch (back) to
adding the alpha identifiers as part of the package transformation step.

Reverted the change to the pkglint rules, and re-ran `yarn pkglint` to update
the package.json files.

I also did some light refactoring, since the method was a bit hard to read.

fixes #15845