chore(assertions): release as an alpha module in CDKv2 (#15614)
The `@aws-cdk/assertions` library was previously excluded from the V2 build of the experimental modules,
because the `vendor.sh` script it used for vendoring in a few modules
(`cfnspec`, `cloudformation-diff`, and `assert-internal`)
used paths relative to the V1 module inside the repository.

Change the `vendor.sh` script to short-circuit and exit when running in `transform` phase for
the alpha packages. That's required because the transform phase re-writes imports when copying
the TS files (because it needs to know which modules are also experimental to do the rewriting correctly),
but the `vendor.sh` script runs only after the `build` script for the package has been invoked,
which means any TS files copied by it successfully would _not_ have their imports re-written.
Since the build for alpha packages runs after the build for `@aws-cdk/assertions`
(we add an explicit dependency to make sure that happens),
those files are already "vendored in" correctly, and will be copied when creating the separate package.

Also adjust the module copying logic to account for bundled dependencies,
which `@aws-cdk/assertion` uses.

Closes #15589

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*