fix(ec2): "clientVpnEndoint" => "clientVpnEndpoint" (#14902)
It is a solution to the issue #13810

Corrected the typo "clientVpnEndpoint" in the ec2 module to "clientVpnEnpoint".

The maintainer (@rix0rrr) talked about @deprecated tags on existing incorrect members in the issue thread, but I didn't work on it because I thought it was an internal reference member. (if @deprecated tag is the rule of this repository in this case, please comment).

This is my first commitment to this report. If there's anything wrong, please let me know in the comments. thanks

---- edit -----
Originally, this PR was #14835 , but I deleted my fork report by mistake, so I couldn't correct the error, so I uploaded it again.