docs(cognito): describe addDependency for identity provider and app client (#15893)
When an app client is created, it receives the name of the identity provider as a string.
During the deployment, the identity provider must be created first, then the app client can be created (because it references the identity provider).
Due to the fact that the reference is passed as a string, CDK can't add the dependency between the identity provider and the app client automatically. Therefore, `addDependency` must be called manually when both constructs are created in the same stack.

The missing dependency was reported two times in the last weeks (issue #15850 and #15692). Because of that I would propose to add this behaviour in the readme.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*