fix(core): asset bundling fails for non-existent user (#15313)
Currently, when not explicitly passing a user in BundlingOptions, `os.userInfo()` is called as a default fallback. Even though the user is only used and needed when using docker bundling, it is still called when using local bundling. This can be problematic e.g. when running cdk scripts inside a docker container where the user does not exist inside the container (see #15415).

Fixes #15415.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*