chore(cdk-release): track and bump separate alpha version
As part of the project to release our alpha modules as independent modules, this
change extends our `cdk-release` tool to support tracking and bumping two
versions: the primary, stable version, and an additional, optional alpha
version.

If the local `version.vX.json` file has an `alphaVersion` property, this version
will be bumped alongside the main (stable) version. If the main version is also
a prerelease, then the alphaVersion is simply incremented (e.g., `2.0.0-alpha.0`
-> `2.0.0-alpha.1`); if the main version is stable, the alpha is incremented to
match (e.g., the `2.1.0` release will get an alpha of `2.1.0-alpha.0`). If no
`alphaVersion` is present, it is simply ignored.

This is still only part of the work to release the alpha modules. Remaining:
- Create the indepedent module CHANGELOGs, and aggregate into the main CHANGELOG
- Change the `align-versions` script to read and apply the alphaVersion if the
  module is an alpha module.
- Create the first alpha version in the v2-main branch.

*Implementation details:*

I was frustrated by the generic 'Updater' and 'UpdaterModule' pattern that I
needed to work around, so I decided to just simplify by removing the generic
capabilities that we're not using. This involved deleting a lot of code that
wasn't doing much. Lots of refactoring ensured. Similarly, there simply weren't
nearly enough tests in this module yet; I added at least a few more to improve
coverage in anticipation of the next tasks (which will require even more tests).
My apologies to the reviewer(s) for the muddled diff, but I strongly believe
these simplifications will make it easier to maintain `cdk-release` going
forward.

related #15591