chore(ecs): allow selecting Bottlerocket via `machineImage` (#16038)
As introduced in #10097, Bottlerocket had to be explicitly (and only) selected
via setting `machineImageType`, which would pick an appropriate
`machineImage`.

Setting `machineImage` to `new BottleRocketImage()` would not be
sufficient, since the feature also requires configuring additional
UserData commands which are only added if `machineImageType` was set.

This method of configuration does not allow customization of the AMI,
such as introduced in #16021.

Instead, we reverse the logic: `machineImageType` may still be necessary
to autoconfigure UserData if we can't know what the machineImage is
(for example in case of a preconfigured AutoScalingGroup), but otherwise
is derived from what `machineImage` is being used.

We allow configuring both fields at the same time for the case when the
autodetection fails.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*