feat(assets): exclude "cdk.out" from docker assets (#16034)
This PR is intended to cover a small bug that creates an infinite loop generating folders in cdk.out if the project root path is specified as the build directory.

With this fix 'cdk.out' folder will be added as glob exclude pattern no matters what.

More details in issue #14841

Original PR from @AlexGue

Closes #14841
Closes #14842



----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*