fix(cli): 'deploy' and 'diff' silently does nothing when given unknown stack name (#16073)
Currently, `cdk deploy` and `cdk diff` on stacks that do not exist return no output on the command line. This PR introduces a descriptive error message for those cases so it is easier to understand what happened.

Leaving the idea of fuzzy matching for stack name suggestions to a (potential) future PR, I can open a new issue for that if this one goes through.

For the linter, I'm not sure if anything of value can be added to the readme so I'm wondering if this would be a candidate for exemption.

closes #15866 while adding the same idea to `diff` as well as `deploy`.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*