chore: added links to the relevant aws cdk workshop tutorials in sample-app readme for the various languages (#16042)
When getting started with the AWS CDK, it is useful to have links to a good tutorial to help learn the tool.
This pr adds links to the generated README.md files in order to access the CDK Workshops.
This is a trivial addition but may help folks who are unaware that the CDK Workshops exist