fix(sqs): unable to import a FIFO queue when the queue ARN is a token (#15976)
The fifo property of a SQS queue is determined based on the queue name. The suffix `.fifo` indicates that it is a fifo queue.
When the queue is imported from a token, the name is not known on synthesis time. Therefore, the name of the queue can't be used to initialize the queue as a FIFO queue.

In some constructs, CDK checks during synthesis if the given queue is a FIFIO queue. Because of that, an additional option is necessary to specify a FIFO queue. 

A new boolean property `fifo` is introduced which can be used to specify a FIFO queue that is imported from a token. The property is optional and is only necessary when a FIFO queue should be imported from a token.

Closes #12466.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*