fix(s3): bucket is not emptied before update when the name changes (#16203)
Changing the bucket name leads CloudFormation to try to delete the bucket and create a new one with the new name. If the bucket is not empty, the deployment will fail. With this change, the custom resource will clear the old bucket when it detects that there has been a name change.

NB: this custom resource is created only when `autoDeleteObjects: true` is passed to the bucket.

Fixes #14011.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*