feat(ecs-patterns): Allow configuration of SSL policy for listeners created by ECS patterns (#15210)
This pr adds the ability to specify the SSL policy of the listener created by an ECS L3 pattern.

The Listener's SSL Policy is immutable once it is created so I felt like this is best way to add this feature.   I plan to follow this PR up with another that implements #11841 and allows further configuration of the listener after it has been created.

closes #8816
----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*